[DEL] crm_rma_priority_extended module is deleted and its functionalities are molted to crm_claim_rma
