+++ crm_claim_rma/crm_claim_rma.py
@@ -83,6 +83,25 @@ class claim_line(models.Model):
         result = seller.get_warranty_return_partner()
         return result
 
+    @api.one
+    @api.depends('guarantee_limit')
+    def _set_priority(self):
+        """
+        To determine the priority of claim line
+        """
+        date_invoice = self.invoice_line_id.invoice_id.date_invoice
+        if self.guarantee_limit and date_invoice:
+            days = fields.datetime.strptime(self.guarantee_limit,
+                                            '%Y-%m-%d') - \
+                fields.datetime.strptime(date_invoice,
+                                         '%Y-%m-%d')
+            if days.days <= 1:
+                self.priority = '3_very_high'
+            elif days.days <= 7:
+                self.priority = '2_high'
+            else:
+                self.priority = '1_normal'
+
     company_id = fields.Many2one(
         'res.company', string='Company', readonly=False,
         change_default=True,
@@ -90,6 +109,16 @@ class claim_line(models.Model):
             'claim.line'))
 
     name = fields.Char('Description', default='none', required=True)
+
+    priority = fields.Selection([('0_not_define', 'Not Define'),
+                                ('1_normal', 'Normal'),
+                                ('2_high', 'High'),
+                                ('3_very_high', 'Very High')],
+                                'Priority', default='0_not_define',
+                                compute='_set_priority',
+                                store=True,
+                                readonly=False,
+                                help="Priority attention of claim line")
 
     claim_condition = fields.\
         Selection([('auth_service', 'Authorized Service'),
@@ -427,6 +456,7 @@ class crm_claim(models.Model):
         change_default=True,
         default=lambda self: self.env['res.company']._company_default_get(
             'crm.claim'))
+
 
     number = fields.Char('Number', readonly=True,
                          required=True,
