--- crm_rma_priority_extended/view/crm_claim.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-
-        <record id="crm_claim_line_search_view" model="ir.ui.view">
-            <field name="name">CRM - Claims Search</field>
-            <field name="model">claim.line</field>
-            <field name="inherit_id" ref="crm_claim_rma.view_crm_claim_lines_filter"/>
-            <field name="arch" type="xml">
-                <group expand="0" string="Group By..." position="inside">
-                    <filter string="Priority"
-                            name="group_by_priority"
-                            domain="[]"
-                            context="{'group_by':'priority'}"
-                            help="Priority"/>
-                </group>
-            </field>
-        </record>
-
-        <record model="ir.actions.act_window" id="crm_claim_rma.act_crm_case_claim_lines">
-            <field name="context">{'search_default_group_by_priority': True}</field>
-        </record>
-
-        <!--
-        <field name="priority_rma" attrs="{'readonly': [('client', '=', True)]}"/>
-        -->
-
-    </data>
-</openerp>
